requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.1
pdfplumber==0.10.0
gTTS==2.4.0
whisper==1.1.10
requests==2.31.0
httpx==0.27.0
weasyprint==62.3
python-multipart==0.0.6
websockets==12.0
torch==2.8.0
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
jinja2==3.1.2
pillow==10.0.0
scikit-learn==1.3.0
email-validator==2.1.1
aiosmtplib==3.0.2
# ATS functionality dependencies
PyPDF2==3.0.1
google-generativeai==0.3.2